Order full process working successfully, permission set for admin and employee

+++ src/main/java/com/ahsan/scrap/model/Product.java
@@ -23,6 +23,6 @@ public class Product {
     @Column(name = "short_name")
 	private String shortName;
     @Column(name = "unit_price")
-    private double unitPrice;
-	private long quantity;
+    private Float unitPrice;
+	private Float quantity;
 }
